requirements.txt
# Function dependencies, for example:
# package>=version
google-cloud-bigquery>=2.10.0
google-cloud-storage>=1.36.1